+++ test/mocha.opts
@@ -1,0 +1,1 @@
+--compilers js:babel-core/register
